// frontend/src/api/models/PageResponseTaskDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TaskDTO } from './TaskDTO';
import {
    TaskDTOFromJSON,
    TaskDTOFromJSONTyped,
    TaskDTOToJSON,
} from './TaskDTO';

/**
 * 
 * @export
 * @interface PageResponseTaskDTO
 */
export interface PageResponseTaskDTO {
    /**
     * 
     * @type {Array<TaskDTO>}
     * @memberof PageResponseTaskDTO
     */
    data?: Array<TaskDTO>;
    /**
     * 
     * @type {number}
     * @memberof PageResponseTaskDTO
     */
    pageNum?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseTaskDTO
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseTaskDTO
     */
    count?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseTaskDTO
     */
    totalPages?: number;
}

/**
 * Check if a given object implements the PageResponseTaskDTO interface.
 */
export function instanceOfPageResponseTaskDTO(value: object): value is PageResponseTaskDTO {
    return true;
}

export function PageResponseTaskDTOFromJSON(json: any): PageResponseTaskDTO {
    return PageResponseTaskDTOFromJSONTyped(json, false);
}

export function PageResponseTaskDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PageResponseTaskDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(TaskDTOFromJSON)),
        'pageNum': json['pageNum'] == null ? undefined : json['pageNum'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'count': json['count'] == null ? undefined : json['count'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
    };
}

export function PageResponseTaskDTOToJSON(value?: PageResponseTaskDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(TaskDTOToJSON)),
        'pageNum': value['pageNum'],
        'pageSize': value['pageSize'],
        'count': value['count'],
        'totalPages': value['totalPages'],
    };
}